refactor(storage): relocate export/storage to /shared for worker reuse (#158)
Move export_stream, serializers and the storage package from services/api
into /shared so the export/import worker (whose image carries only
services/workers/* + /shared) can import the same streaming generators and
ObjectStorageService singleton. Old API paths become thin sys.modules-alias
shims so existing call sites and mock.patch targets resolve to the same module
object (singleton identity preserved).

Reconcile the storage env vars with storage_config.py and the Helm values:
the singleton now reads STORAGE_TYPE / S3_ACCESS_KEY / S3_SECRET_KEY as the
canonical names (legacy STORAGE_BACKEND / *_ID / AWS_* kept as fallbacks),
honours S3_USE_SSL, and signs browser-facing presigned URLs against a second
client bound to STORAGE_BASE_URL (SigV4 binds the request host, so an
internally-signed URL would 403 when replayed against the public host).

Add boto3/botocore to api pyproject.toml (the image's actual install source)
and workers requirements.txt; without them the singleton silently falls back
to local filesystem even when STORAGE_TYPE=minio.

Behavior is unchanged when STORAGE_TYPE is unset (default-OFF: local backend).
