+++ services/api/storage_service.py
@@ -1,2 +1,10 @@
-"""Compatibility shim -- real implementation in services/storage/storage_service.py"""
-from services.storage.storage_service import *  # noqa: F401, F403
+"""Compatibility shim — implementation moved to /shared/storage/storage_service.py (issue #158).
+
+Aliases the top-level ``storage_service`` name to the canonical ``storage.storage_service``
+module. /shared is first on sys.path in both the api and worker containers.
+"""
+import sys
+
+import storage.storage_service as _canonical
+
+sys.modules[__name__] = _canonical
