+++ services/workers/requirements.txt
@@ -22,6 +22,11 @@ rouge-score>=0.1.2
 sacrebleu>=2.4.0
 jsonschema==4.26.0
 ijson>=3.2.0  # Streaming JSON parser for memory-bounded import worker (issue #158)
+# boto3/botocore: the export/import worker streams artifacts to S3-compatible
+# object storage (multipart upload / download); without them the relocated
+# storage singleton falls back to local filesystem (issue #158).
+boto3>=1.35.0
+botocore>=1.43.11
 pytest==8.4.2
 pytest-asyncio==0.26.0
 # AI service dependencies
